# public/SchoolManSystemOpenAPI.yaml
openapi: 3.0.3
info:
  title: School Management System
  description: >-
    # 🚀 Get started here


    This template guides you through CRUD operations (GET, POST, PUT, DELETE),
    variables, and tests.


    ## 🔖 **How to use this template**


    #### **Step 1: Send requests**


    RESTful APIs allow you to perform CRUD operations using the POST, GET, PUT,
    and DELETE HTTP methods.


    This collection contains each of these
    [request](https://learning.postman.com/docs/sending-requests/requests/)
    types. Open each request and click "Send" to see what happens.


    #### **Step 2: View responses**


    Observe the response tab for status code (200 OK), response time, and size.


    #### **Step 3: Send new Body data**


    Update or add new data in "Body" in the POST request. Typically, Body data
    is also used in PUT request.


    ```

    {
        "name": "Add your name in the body"
    }

     ```

    #### **Step 4: Update the variable**


    Variables enable you to store and reuse values in Postman. We have created a
    [variable](https://learning.postman.com/docs/sending-requests/variables/)
    called `base_url` with the sample request
    [https://postman-api-learner.glitch.me](https://postman-api-learner.glitch.me).
    Replace it with your API endpoint to customize this collection.


    #### **Step 5: Add tests in the "Tests" tab**


    Tests help you confirm that your API is working as expected. You can write
...


    <img
...


    ## 💪 Pro tips

...
servers:
...
paths:
...
  /api/class/3:
    get:
      tags:
        - Class
      summary: Get Class By Id
      description: Get Class By Id
      operationId: getClassById
      requestBody:
        content:
          text/plain:
            example: ''
      responses:
        '200':
          description: ''
    delete:
      tags:
        - Class
      summary: Delete Class
      description: Delete Class
      operationId: deleteClass
      requestBody:
        content:
          text/plain:
            example: ''
      responses:
        '200':
          description: ''
  /api/class/1/teacher:
    get:
      tags:
        - Class
      summary: Get Teachers of Class
      description: Get Teachers of Class
      operationId: getTeachersOfClass
      requestBody:
        content:
          text/plain:
            example: ''
      responses:
        '200':
          description: ''
  /api/class/1/student:
    get:
      tags:
        - Class
      summary: Get Students of Class
      description: Get Students of Class
      operationId: getStudentsOfClass
      requestBody:
        content:
          text/plain:
            example: ''
      responses:
        '200':
          description: ''
...
  /api/class/1/teacher/1:
    post:
      tags:
        - Class
      summary: Add Teacher To Class
      description: Add Teacher To Class
      operationId: addTeacherToClass
      requestBody:
        content:
          text/plain:
            example: ''
      responses:
        '200':
          description: ''
  /api/class/2/student/2:
    post:
      tags:
        - Class
      summary: Add Student To Class
      description: Add Student To Class
      operationId: addStudentToClass
      requestBody:
        content:
          text/plain:
            example: ''
      responses:
        '200':
          description: ''
    delete:
      tags:
        - Class
      summary: Remove Student from Class
      description: Remove Student from Class
      operationId: removeStudentFromClass
      requestBody:
        content:
          text/plain:
            example: ''
      responses:
        '200':
          description: ''
  /api/class/1/teacher/2:
    delete:
      tags:
        - Class
      summary: Remove Teacher from Class
      description: Remove Teacher from Class
      operationId: removeTeacherFromClass
      requestBody:
        content:
          text/plain:
            example: ''
      responses:
        '200':
          description: ''
  /api/class/2:
    put:
      tags:
        - Class
      summary: Update Class
      description: Update Class
      operationId: updateClass
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                code:
                  type: string
                  example: SW101
                counselor:
                  type: string
                  example: Joseph Fox
                name:
                  type: string
                  example: Software Engineering
            example:
              code: SW101
              counselor: Joseph Fox
              name: Software Engineering
      responses:
        '200':
          description: ''
  /api/teacher:
    get:
      tags:
        - Teacher
      summary: Get All Teachers
      description: Get All Teachers
      operationId: getAllTeachers
      requestBody:
        content:
          text/plain:
            example: ''
      responses:
        '200':
          description: ''
...
  /api/teacher/1:
    get:
      tags:
        - Teacher
      summary: Get Classes of a Teacher
      description: Get Classes of a Teacher
      operationId: getClassesOfATeacher
      responses:
        '200':
          description: ''
    put:
      tags:
        - Teacher
      summary: Update Teacher
      description: Update Teacher
      operationId: updateTeacher
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: ''
                expertise:
                  type: string
                  example: ''
                name:
                  type: string
                  example: ''
                password:
                  type: string
                  example: admin
                role:
                  type: string
                  example: teacher
            example:
              email: ''
              expertise: ''
              name: ''
              password: admin
              role: teacher
      responses:
        '200':
          description: ''
    delete:
      tags:
        - Teacher
      summary: Delete Teacher
      description: Delete Teacher
      operationId: deleteTeacher
      responses:
        '200':
          description: ''
  /api/teacher/1/student:
    get:
      tags:
        - Teacher
      summary: Get Students of Teacher
      description: Get Students of Teacher
      operationId: getStudentsOfTeacher
      requestBody:
        content:
          text/plain:
            example: ''
      responses:
        '200':
          description: ''
...
  /api/class-schedule:
    post:
      tags:
        - Class Schedule
      summary: Create Class Schedule
      description: Create Class Schedule
      operationId: createClassSchedule
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                daysOfWeek:
                  type: string
                  example: 1,3,5
                endDate:
                  type: string
                  example: '2024-02-31'
                startDate:
                  type: string
                  example: '2024-02-19'
                title:
                  type: string
                  example: Reeact Session
            example:
              daysOfWeek: 1,3,5
              endDate: '2024-02-31'
              startDate: '2024-02-19'
              title: Reeact Session
      responses:
        '200':
          description: ''
  /api/class-schedule/1:
    put:
      tags:
        - Class Schedule
      summary: Update Class Schedule
      description: Update Class Schedule
      operationId: updateClassSchedule
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                daysOfWeek:
                  type: string
                  example: '1'
                endDate:
                  type: string
                  example: '2024-04-15T22:30:00'
                startDate:
                  type: string
                  example: '2024-02-19T19:30:00'
                title:
                  type: string
                  example: Bootcamp
            example:
              daysOfWeek: '1'
              endDate: '2024-04-15T22:30:00'
              startDate: '2024-02-19T19:30:00'
              title: Bootcamp
      responses:
        '200':
          description: ''
...
  /api/file/upload:
    post:
      tags:
        - File Handling
      summary: Upload File
      description: Upload File
      operationId: uploadFile
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                classIds:
                  type: string
                  example: '1'
                file:
                  type: string
                  format: binary
                isPublic:
                  type: string
                  example: 'true'
      responses:
        '200':
          description: ''
  /api/file/download/file-94e96fff-ee46-4706-bd12-be9bdc77eb4e.png:
    get:
      tags:
        - File Handling
      summary: Download File
      description: Download File
      operationId: downloadFile
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: ''
...